+++ lib/src/flutter_recorder_plugin.dart
